Moved a couple of functions in the proper builder package
