+++ legacy/builder/create_cmake_rule.go
@@ -119,7 +119,7 @@ func (s *ExportProjectCMake) Run(ctx *types.Context) error {
 	}
 
 	// Use old ctags method to generate export file
-	ctx.SketchSourceMerged = FilterSketchSource(ctx.Sketch, strings.NewReader(ctx.SketchSourceMerged), true)
+	ctx.SketchSourceMerged = filterSketchSource(ctx.Sketch, strings.NewReader(ctx.SketchSourceMerged), true)
 
 	err = utils.CopyDir(ctx.SketchBuildPath.String(), cmakeFolder.Join("sketch").String(), validExportExtensions)
 	if err != nil {
