+++ legacy/builder/test/utils_test.go
@@ -65,41 +65,3 @@ func TestQuoteCppString(t *testing.T) {
 		require.Equal(t, expected, utils.QuoteCppString(input))
 	}
 }
-
-func TestParseCppString(t *testing.T) {
-	_, _, ok := utils.ParseCppString(`foo`)
-	require.Equal(t, false, ok)
-
-	_, _, ok = utils.ParseCppString(`"foo`)
-	require.Equal(t, false, ok)
-
-	str, rest, ok := utils.ParseCppString(`"foo"`)
-	require.Equal(t, true, ok)
-	require.Equal(t, `foo`, str)
-	require.Equal(t, ``, rest)
-
-	str, rest, ok = utils.ParseCppString(`"foo\\bar"`)
-	require.Equal(t, true, ok)
-	require.Equal(t, `foo\bar`, str)
-	require.Equal(t, ``, rest)
-
-	str, rest, ok = utils.ParseCppString(`"foo \"is\" quoted and \\\\bar\"\" escaped\\" and "then" some`)
-	require.Equal(t, true, ok)
-	require.Equal(t, `foo "is" quoted and \\bar"" escaped\`, str)
-	require.Equal(t, ` and "then" some`, rest)
-
-	str, rest, ok = utils.ParseCppString(`" !\"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_abcdefghijklmnopqrstuvwxyz{|}~"`)
-	require.Equal(t, true, ok)
-	require.Equal(t, ` !"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\]^_abcdefghijklmnopqrstuvwxyz{|}~`, str)
-	require.Equal(t, ``, rest)
-
-	str, rest, ok = utils.ParseCppString(`"/home/ççç/"`)
-	require.Equal(t, true, ok)
-	require.Equal(t, `/home/ççç/`, str)
-	require.Equal(t, ``, rest)
-
-	str, rest, ok = utils.ParseCppString(`"/home/ççç/ /$sdsdd\\"`)
-	require.Equal(t, true, ok)
-	require.Equal(t, `/home/ççç/ /$sdsdd\`, str)
-	require.Equal(t, ``, rest)
-}
